__print_end(): fix printf(3) warnings
In the previous implementation, the desired output string was directly
used in place of the format string fed to printf(3). This caused the
interpretation of some parts of the ASCII art as special characters
which in turn triggered compilation warnings.

A non escaped '\' was also fixed in the desired output string.